finished base tetris game
