+++ src/main/Main.java
@@ -1,0 +1,20 @@
+package main;
+
+import javax.swing.JFrame;
+
+public class Main {
+    public static void main(String[] args) {
+        JFrame window = new JFrame("Sir Tet");
+        window.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        window.setResizable(false);
+
+        GamePanel gp = new GamePanel();
+        window.add(gp);
+        window.pack();
+
+        window.setLocationRelativeTo(null);
+        window.setVisible(true);
+
+        gp.LaunchGame();
+    }
+}
