+++ src/mino/Mino.java
@@ -1,0 +1,230 @@
+package mino;
+
+import java.awt.Color;
+import java.awt.Graphics2D;
+
+import main.KeyHandler;
+import main.PlayManager;
+
+public class Mino {
+    public Block b[] = new Block[4];
+    public Block tempB[] = new Block[4];
+    int autoDropCounter = 0;
+    public int direction = 1;
+    boolean leftCollision, rightCollision, bottomCollision;
+    public boolean active = true;
+    public boolean deactivating;
+    int deactivateCounter = 0;
+
+    public void create(Color c) {
+        b[0] = new Block(c);
+        b[1] = new Block(c);
+        b[2] = new Block(c);
+        b[3] = new Block(c);
+        tempB[0] = new Block(c);
+        tempB[1] = new Block(c);
+        tempB[2] = new Block(c);
+        tempB[3] = new Block(c);
+    }
+
+    public void setXY(int x, int y) {
+
+    }
+
+    public void updateXY(int direction) {
+
+        checkRotationCollision();
+        if (!leftCollision && !rightCollision && !bottomCollision) {
+            this.direction = direction;
+            b[0].x = tempB[0].x;
+            b[0].y = tempB[0].y;
+            b[1].x = tempB[1].x;
+            b[1].y = tempB[1].y;
+            b[2].x = tempB[2].x;
+            b[2].y = tempB[2].y;
+            b[3].x = tempB[3].x;
+            b[3].y = tempB[3].y;
+        }
+    }
+
+    public void getDirection1() {
+
+    }
+
+    public void getDirection2() {
+        
+    }
+
+    public void getDirection3() {
+        
+    }
+
+    public void getDirection4() {
+        
+    }
+
+    public void checkMovementCollision() {
+        leftCollision = false;
+        rightCollision = false;
+        bottomCollision = false;
+
+        checkStaticBlockCollision();
+
+        for (Block b1 : b) {
+            if (b1.x == PlayManager.left_x) {
+                leftCollision = true;
+            }
+        }
+
+        for (Block b1 : b) {
+            if (b1.x + Block.SIZE == PlayManager.right_x) {
+                rightCollision = true;
+            }
+        }
+
+        for (Block b1 : b) {
+            if (b1.y + Block.SIZE == PlayManager.bottom_y) {
+                bottomCollision = true;
+            }
+        }
+    }
+
+    public void checkRotationCollision() {
+        leftCollision = false;
+        rightCollision = false;
+        bottomCollision = false;
+
+        checkStaticBlockCollision();
+
+        for (Block b1 : tempB) {
+            if (b1.x < PlayManager.left_x) {
+                leftCollision = true;
+            }
+        }
+
+        for (Block b1 : tempB) {
+            if (b1.x + Block.SIZE > PlayManager.right_x) {
+                rightCollision = true;
+            }
+        }
+
+        for (Block b1 : tempB) {
+            if (b1.y + Block.SIZE > PlayManager.bottom_y) {
+                bottomCollision = true;
+            }
+        }
+    }
+
+    private void checkStaticBlockCollision() {
+        for (int i = 0; i < PlayManager.staticBlocks.size(); i++) {
+            int targetX = PlayManager.staticBlocks.get(i).x;
+            int targetY = PlayManager.staticBlocks.get(i).y;
+            for (Block b1 : b) {
+                if (b1.y + Block.SIZE == targetY && b1.x == targetX) {
+                    bottomCollision = true;
+                }
+            }
+            for (Block b1 : b) {
+                if (b1.x - Block.SIZE == targetX && b1.y == targetY) {
+                    leftCollision = true;
+                }
+            }
+            for (Block b1 : b) {
+                if (b1.x + Block.SIZE == targetX && b1.y == targetY) {
+                    rightCollision = true;
+                }
+            }
+        }
+    }
+
+    public void update() {
+
+        if (deactivating) {
+            deactivating();
+        }
+
+        if (KeyHandler.upPressed) {
+            switch (direction) {
+                case 1 -> getDirection2();
+                case 2 -> getDirection3();
+                case 3 -> getDirection4();
+                case 4 -> getDirection1();
+            }
+            KeyHandler.upPressed = false;
+        }
+
+        checkMovementCollision();
+
+        if (KeyHandler.downPressed) {
+            if (bottomCollision == false) {
+                b[0].y += Block.SIZE;
+                b[1].y += Block.SIZE;
+                b[2].y += Block.SIZE;
+                b[3].y += Block.SIZE;
+            }
+
+            autoDropCounter = 0;
+            KeyHandler.downPressed = false;
+        }
+
+        if (KeyHandler.leftPressed) {
+            if (leftCollision == false) {
+                b[0].x -= Block.SIZE;
+                b[1].x -= Block.SIZE;
+                b[2].x -= Block.SIZE;
+                b[3].x -= Block.SIZE;
+            }
+
+            KeyHandler.leftPressed = false;
+        }
+
+        if (KeyHandler.rightPressed) {
+            if (rightCollision == false) {
+                b[0].x += Block.SIZE;
+                b[1].x += Block.SIZE;
+                b[2].x += Block.SIZE;
+                b[3].x += Block.SIZE;
+            }
+
+            KeyHandler.rightPressed = false;
+        }
+
+        if (bottomCollision) {
+            deactivating = true;
+        }
+        else {
+            autoDropCounter++;
+            if (autoDropCounter == PlayManager.dropInterval) {
+                b[0].y += Block.SIZE;
+                b[1].y += Block.SIZE;
+                b[2].y += Block.SIZE;
+                b[3].y += Block.SIZE;
+                autoDropCounter = 0;
+            }
+        }
+
+    }
+
+    private void deactivating() {
+
+        deactivateCounter++;
+
+        if (deactivateCounter == 45) {
+            deactivateCounter = 0;
+            checkMovementCollision();
+
+            if (bottomCollision) {
+                active = false;
+            }
+        }
+    }
+
+    public void draw(Graphics2D g2) {
+        int margin = 2;
+        g2.setColor(b[0].c);
+        g2.fillRect(b[0].x+margin, b[0].y+margin, Block.SIZE-(margin*2), Block.SIZE-(margin*2));
+        g2.fillRect(b[1].x+margin, b[1].y+margin, Block.SIZE-(margin*2), Block.SIZE-(margin*2));
+        g2.fillRect(b[2].x+margin, b[2].y+margin, Block.SIZE-(margin*2), Block.SIZE-(margin*2));
+        g2.fillRect(b[3].x+margin, b[3].y+margin, Block.SIZE-(margin*2), Block.SIZE-(margin*2));
+    }
+}
